Fixed importer

+++ test/data/import.js
@@ -101,20 +101,30 @@ let lib = {
 			records.push(rec);
 		});
 		if (records && Array.isArray(records) && records.length > 0) {
-			async.each(
+			let tenants = [];
+			async.eachSeries(
 				records,
 				(e, cb) => {
 					profile.name = e.tenant.code + "_urac";
 					let mongoConnection = new Mongo(profile);
-					mongoConnection.dropCollection("users", () => {
+					if (tenants.includes(profile.name)) {
 						let condition = {email: e.email};
 						e._id = mongoConnection.ObjectId(e._id);
 						mongoConnection.update("users", condition, e, {'upsert': true}, (error, result) => {
-							console.log("users", error);
 							mongoConnection.closeDb();
 							return cb();
 						});
-					});
+					} else {
+						tenants.push(profile.name);
+						mongoConnection.dropCollection("users", () => {
+							let condition = {email: e.email};
+							e._id = mongoConnection.ObjectId(e._id);
+							mongoConnection.update("users", condition, e, {'upsert': true}, (error, result) => {
+								mongoConnection.closeDb();
+								return cb();
+							});
+						});
+					}
 				},
 				() => {
 					return cb();
@@ -130,12 +140,13 @@ let lib = {
 			records.push(rec);
 		});
 		if (records && Array.isArray(records) && records.length > 0) {
-			async.each(
+			let tenants = [];
+			async.eachSeries(
 				records,
 				(e, cb) => {
 					profile.name = e.tenant.code + "_urac";
 					let mongoConnection = new Mongo(profile);
-					mongoConnection.dropCollection("groups", () => {
+					if (tenants.includes(profile.name)) {
 						let condition = {code: e.code};
 						e._id = mongoConnection.ObjectId(e._id);
 						mongoConnection.update("groups", condition, e, {'upsert': true}, (error, result) => {
@@ -143,7 +154,18 @@ let lib = {
 							mongoConnection.closeDb();
 							return cb();
 						});
-					});
+					} else {
+						tenants.push(profile.name);
+						mongoConnection.dropCollection("groups", () => {
+							let condition = {code: e.code};
+							e._id = mongoConnection.ObjectId(e._id);
+							mongoConnection.update("groups", condition, e, {'upsert': true}, (error, result) => {
+								console.log("groups", error);
+								mongoConnection.closeDb();
+								return cb();
+							});
+						});
+					}
 				},
 				() => {
 					return cb();
